organize tests to colins suggestion
